Move preprocess effects to util file, merge JSON object & array effects.
