+++ app/src/app/api/reactions/reversible/schemas.ts
@@ -2,16 +2,16 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { queryJSONArraySchema } from "@/docs/openapi";
 import { ReactionTypeTag } from "@lxcat/schema/process";
 import { z } from "zod";
 import { stateLeafSchema } from "../../schemas.openapi";
+import { queryJSONSchema } from "../../util";
 
 export const querySchema = z.object({
   query: z.object({
-    consumes: queryJSONArraySchema(z.array(stateLeafSchema)),
-    produces: queryJSONArraySchema(z.array(stateLeafSchema)),
-    typeTags: queryJSONArraySchema(z.array(ReactionTypeTag)),
-    setIds: queryJSONArraySchema(z.array(z.string())),
+    consumes: queryJSONSchema(z.array(stateLeafSchema)),
+    produces: queryJSONSchema(z.array(stateLeafSchema)),
+    typeTags: queryJSONSchema(z.array(ReactionTypeTag)),
+    setIds: queryJSONSchema(z.array(z.string())),
   }),
 });
