+++ app/src/app/api/reactions/type-tags/schemas.ts
@@ -2,16 +2,16 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { queryJSONArraySchema } from "@/docs/openapi";
 import { Reversible } from "@lxcat/database/item/picker";
 import { z } from "zod";
 import { stateLeafSchema } from "../../schemas.openapi";
+import { queryJSONSchema } from "../../util";
 
 export const querySchema = z.object({
   query: z.object({
-    consumes: queryJSONArraySchema(z.array(stateLeafSchema)),
-    produces: queryJSONArraySchema(z.array(stateLeafSchema)),
+    consumes: queryJSONSchema(z.array(stateLeafSchema)),
+    produces: queryJSONSchema(z.array(stateLeafSchema)),
     reversible: z.nativeEnum(Reversible).default(Reversible.Both),
-    setIds: queryJSONArraySchema(z.array(z.string())),
+    setIds: queryJSONSchema(z.array(z.string())),
   }),
 });
