+++ app/src/app/api/scat-cs/inspect/schemas.ts
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { IdsSchema } from "@/app/scat-cs/IdsSchema";
-import { queryArraySchema } from "@/docs/openapi";
 import { z } from "zod";
+import { queryArraySchema } from "../../util";
 
 export const querySchema = z.object({
   query: z.object({
